feat(data structure): change to accept arrays of schema for a single page
Some pages are different from another even on the same website. We don't have control on what they
do. But atleast we can figure out what needs to be scanned. In future perhaps some sort of ML

BREAKING CHANGE: Schemas now needs to be an array instead of a single object.